Fix up HOT seaflow upload

+++ app/upload.sh
@@ -5,8 +5,10 @@
 date
 ./mc.darwin cp -q \
     ~/Desktop/KM_underway/HOT_{GPGGA,GPVTG,flor,met,uthsl}_feed.txt \
-    minio/input-data/hot_2019_06_30
+    minio/input-data/hot_2019_06_30/
 
-./mc.darwin cp -q \
-	~/realtime/seaflow-realtime-results/sync/stats.csv \
-	minio/input-data/gradients3/KM.seaflow-740.stats.csv
+if [[ -e ~/realtime/seaflow-realtime-results/sync/stats.csv ]]; then
+    ./mc.darwin cp -q \
+        ~/realtime/seaflow-realtime-results/sync/stats.csv \
+        minio/input-data/hot_2019_06_30/HOT.seaflow-740.stats.csv
+fi
